requirements.txt
python-dotenv
openai
pydantic
requests